Move node and message handling to LoginRepository
